server/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
b2sdk==2.8.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cors==1.0.1
cryptography==41.0.3
Faker==19.8.0
filelock==3.17.0
Flask==2.3.2
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
logfury==1.0.1
Mako==1.3.8
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==1.24.3
packaging==24.2
pandas==2.1.1
pycparser==2.22
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
requests-file==2.1.0
six==1.17.0
SQLAlchemy==2.0.37
tldextract==5.1.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
